refactor(storage): refactor code related to LocalVersion to make the code cleaner
