+++ src/storage/src/hummock/iterator/forward_user.rs
@@ -22,7 +22,7 @@ use crate::hummock::iterator::{
     BackwardUserIterator, ConcatIteratorInner, Forward, HummockIterator, HummockIteratorDirection,
     HummockIteratorUnion,
 };
-use crate::hummock::local_version::PinnedVersion;
+use crate::hummock::local_version::pinned_version::PinnedVersion;
 use crate::hummock::shared_buffer::shared_buffer_batch::SharedBufferBatchIterator;
 use crate::hummock::shared_buffer::SharedBufferIteratorType;
 use crate::hummock::value::HummockValue;
